Major Update.
Preparing to merge with development.
Removed AWS related files.
Removed AWS related imports.
Removed AWS related dependencies.
