--- api/internal/interfaces/rest-firestore/middleware/rate_limiter.go
@@ -1,22 +1,0 @@
-package firestore_middleware
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-	"golang.org/x/time/rate"
-)
-
-var limiter = rate.NewLimiter(1, 30)
-
-func RateLimiter() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		if !limiter.Allow() {
-			c.AbortWithStatusJSON(http.StatusTooManyRequests, gin.H{
-				"error": "Too Many Requests!",
-			})
-			return
-		}
-		c.Next()
-	}
-}
