--- api/internal/interfaces/rest-firestore/server.go
@@ -1,15 +1,0 @@
-package firestore_server
-
-import (
-	"github.com/gin-gonic/gin"
-)
-
-func Server() {
-	gin.SetMode(gin.DebugMode)
-
-	r := gin.Default()
-
-	RegisteredRoutes(r)
-
-	r.Run()
-}
