+++ api/internal/interfaces/rest/handlers/chapters.go
