+++ api/internal/interfaces/rest/middleware/is_admin.go
@@ -1,4 +1,4 @@
-package middleware
+package firestore_middleware
 
 import (
 	"Codex-Backend/api/internal/domain"
@@ -9,7 +9,7 @@ import (
 
 func IsAdmin() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		var user domain.User
+		var user *domain.User
 
 		// Get user from context
 		result, ok := c.Get("user")
@@ -21,15 +21,15 @@ func IsAdmin() gin.HandlerFunc {
 		}
 
 		// Cast user to User struct
-		user, ok = result.(domain.User)
+		user, ok = result.(*domain.User)
 		if !ok {
-			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-				"error": "Type assertion failed",
+			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{
+				"error": "Incorrect User struct",
 			})
 			return
 		}
 
-		if user.Type != "admin" {
+		if user.Type != "Admin" {
 			c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{
 				"error": "Unauthorized access",
 			})
