--- api/internal/usecases/auth/password.go
@@ -1,11 +1,0 @@
-package auth_service
-
-import "golang.org/x/crypto/bcrypt"
-
-func HashPassword(password string) ([]byte, error) {
-	return bcrypt.GenerateFromPassword([]byte(password), 14)
-}
-
-func VerifyPassword(hashedPassword, password string) error {
-	return bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password))
-}
